requirements.txt
pytorch-lightning>=2.0,<3
transformers>=4.40
peft>=0.10
safetensors
einops
omegaconf
# SMPL pickles (smplx) need chumpy; PyPI chumpy only works with numpy<1.24
numpy>=1.23,<1.24
scipy
tqdm
rich
spacy
Pillow
regex
opencv-python
pyrender
trimesh
mapbox-earcut
smplx
chumpy
shapely
requests
accelerate
fastapi>=0.115,<1
uvicorn[standard]>=0.27
pydantic>=2
huggingface-hub>=0.24,<1
moviepy>=1.0.3,<2
imageio
matplotlib